Fix tray context menu on Linux
