+++ package.json
@@ -167,6 +167,7 @@
     "babel-register": "^6.18.0",
     "buttercup": "~1.5.1",
     "buttercup-importer": "^0.11.0",
+    "check-os-tray-support": "^1.0.0",
     "classnames": "^2.2.5",
     "concurrently": "^3.5.1",
     "cross-env": "^5.1.1",
