+++ src/main/actions.js
@@ -2,7 +2,8 @@ import debounce from 'lodash/debounce';
 import { ipcMain as ipc, BrowserWindow, app } from 'electron';
 import { getWindowManager } from './lib/window-manager';
 import { openFile, newFile, openFileForImporting } from './lib/files';
-import { setupMenu, setupTrayIcon } from './menu';
+import { setupMenu } from './menu';
+import { setupTrayIcon } from './tray';
 import { getMainWindow } from './utils/window';
 import i18n, { languages } from '../shared/i18n';
 import localesConfig from '../../locales/config';
