+++ src/main/app.js
@@ -3,7 +3,8 @@ import pify from 'pify';
 import log from 'electron-log';
 import jsonStorage from 'electron-json-storage';
 import configureStore from '../shared/store/configure-store';
-import { setupMenu, setupTrayIcon } from './menu';
+import { setupMenu } from './menu';
+import { setupTrayIcon } from './tray';
 import { getWindowManager } from './lib/window-manager';
 import { sendEventToMainWindow } from './utils/window';
 import { loadFile } from './lib/files';
