+++ src/main/menu.js
@@ -1,5 +1,5 @@
-import { app, shell, Menu, Tray } from 'electron';
-import { isOSX, isWindows } from '../shared/utils/platform';
+import { app, shell, Menu } from 'electron';
+import { isOSX } from '../shared/utils/platform';
 import {
   getCurrentArchiveId,
   getAllArchives,
@@ -11,105 +11,15 @@ import { openFile, openFileForImporting, newFile } from './lib/files';
 import { toggleArchiveSearch } from './lib/archive-search';
 import { getWindowManager } from './lib/window-manager';
 import { checkForUpdates } from './lib/updater';
-import {
-  getMainWindow,
-  reopenMainWindow,
-  checkDockVisibility
-} from './utils/window';
+import { getMainWindow, reopenMainWindow } from './utils/window';
 import i18n, { languages } from '../shared/i18n';
 import pkg from '../../package.json';
 import electronContextMenu from 'electron-context-menu';
-import { getPathToFile } from './lib/utils';
+import { setupTrayIcon } from './tray';
 
 electronContextMenu();
 
 const label = (key, options) => i18n.t(`app-menu.${key}`, options);
-
-let tray = null;
-let isTrayIconInitialized = false;
-
-export const setupTrayIcon = store => {
-  const state = store.getState();
-  const isTrayIconEnabled = getSetting(state, 'isTrayIconEnabled');
-
-  if (isTrayIconEnabled) {
-    if (!tray) {
-      let trayPath = 'resources/icons/trayTemplate.png';
-      if (isWindows()) {
-        trayPath = 'resources/icons/tray.ico';
-      }
-
-      tray = new Tray(getPathToFile(trayPath));
-    }
-
-    const showTrayMenu = () => {
-      tray.popUpContextMenu(
-        Menu.buildFromTemplate([
-          {
-            label: label('tray.open'),
-            click: () => {
-              reopenMainWindow();
-            }
-          },
-          {
-            type: 'separator'
-          },
-          {
-            label: label('archive.new'),
-            click: () => {
-              reopenMainWindow(win => newFile(win));
-            }
-          },
-          {
-            label: label('archive.open'),
-            click: () => {
-              reopenMainWindow(win => openFile(win));
-            }
-          },
-          {
-            label: label('archive.connect-cloud-sources'),
-            click: () => {
-              reopenMainWindow(win => {
-                getWindowManager().buildWindowOfType('file-manager', null, {
-                  parent: win
-                });
-              });
-            }
-          },
-          {
-            type: 'separator'
-          },
-          {
-            label: label('tray.quit'),
-            click: () => {
-              app.quit();
-            }
-          }
-        ])
-      );
-    };
-
-    if (!isTrayIconInitialized) {
-      tray.on('click', () => {
-        showTrayMenu();
-      });
-
-      tray.on('right-click', () => {
-        showTrayMenu();
-      });
-    }
-
-    isTrayIconInitialized = true;
-  } else {
-    if (tray) {
-      tray.destroy();
-    }
-    tray = null;
-    isTrayIconInitialized = false;
-
-    checkDockVisibility();
-  }
-};
 
 export const setupMenu = store => {
   const defaultTemplate = [
