+++ src/main/tray.js
@@ -1,0 +1,97 @@
+import { app, Menu, Tray } from 'electron';
+import checkTraySupport from 'check-os-tray-support';
+import { isWindows, isLinux } from '../shared/utils/platform';
+import { reopenMainWindow, checkDockVisibility } from './utils/window';
+import { getWindowManager } from './lib/window-manager';
+import { openFile, newFile } from './lib/files';
+import { getSetting } from '../shared/selectors';
+import i18n from '../shared/i18n';
+import { getPathToFile } from './lib/utils';
+
+const label = (key, options) => i18n.t(`app-menu.${key}`, options);
+
+let tray = null;
+let isTrayIconInitialized = false;
+
+export const setupTrayIcon = async store => {
+  const state = store.getState();
+  const isTrayIconEnabled = getSetting(state, 'isTrayIconEnabled');
+
+  if (!await checkTraySupport() || !isTrayIconEnabled) {
+    if (tray) {
+      tray.destroy();
+    }
+    tray = null;
+    isTrayIconInitialized = false;
+
+    checkDockVisibility();
+    return;
+  }
+
+  if (!tray) {
+    let trayPath = 'resources/icons/trayTemplate.png';
+    if (isWindows()) {
+      trayPath = 'resources/icons/tray.ico';
+    }
+
+    tray = new Tray(getPathToFile(trayPath));
+  }
+
+  const contextMenu = Menu.buildFromTemplate([
+    {
+      label: label('tray.open'),
+      click: () => {
+        reopenMainWindow();
+      }
+    },
+    {
+      type: 'separator'
+    },
+    {
+      label: label('archive.new'),
+      click: () => {
+        reopenMainWindow(win => newFile(win));
+      }
+    },
+    {
+      label: label('archive.open'),
+      click: () => {
+        reopenMainWindow(win => openFile(win));
+      }
+    },
+    {
+      label: label('archive.connect-cloud-sources'),
+      click: () => {
+        reopenMainWindow(win => {
+          getWindowManager().buildWindowOfType('file-manager', null, {
+            parent: win
+          });
+        });
+      }
+    },
+    {
+      type: 'separator'
+    },
+    {
+      label: label('app.quit'),
+      click: () => {
+        app.quit();
+      }
+    }
+  ]);
+
+  const showTrayMenu = () => {
+    tray.popUpContextMenu(contextMenu);
+  };
+
+  if (!isTrayIconInitialized) {
+    if (isLinux()) {
+      tray.setContextMenu(contextMenu);
+    } else {
+      tray.on('click', () => reopenMainWindow());
+      tray.on('right-click', () => showTrayMenu());
+    }
+  }
+
+  isTrayIconInitialized = true;
+};
